Spring boot: javadocs for benchmark application.properties + less verbose test logging
