+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/src/main/java/org/optaplanner/spring/boot/autoconfigure/OptaPlannerAutoConfiguration.java
@@ -40,6 +40,10 @@ import org.optaplanner.core.config.solver.SolverConfig;
 import org.optaplanner.core.config.solver.SolverManagerConfig;
 import org.optaplanner.core.config.solver.termination.TerminationConfig;
 import org.optaplanner.persistence.jackson.api.OptaPlannerJacksonModule;
+import org.optaplanner.spring.boot.autoconfigure.config.OptaPlannerProperties;
+import org.optaplanner.spring.boot.autoconfigure.config.SolverManagerProperties;
+import org.optaplanner.spring.boot.autoconfigure.config.SolverProperties;
+import org.optaplanner.spring.boot.autoconfigure.config.TerminationProperties;
 import org.optaplanner.test.api.score.stream.ConstraintVerifier;
 import org.optaplanner.test.api.score.stream.MultiConstraintVerification;
 import org.optaplanner.test.api.score.stream.SingleConstraintVerification;
